fix(audit): restore 8 merge regressions from Batch I structural split
- B-2: persecution intensity formula back to minority_ratio (not 1-ratio)
- H-21: HOLY_WAR_WEIGHT_BONUS restored to 1.75 (multiplicative)
- H-7: capital_region=None on 3 absorption paths
- H-8: dead-civ guards in collect_tribute and check_federation_formation
- C-9: FFI unwrap replaced with map_err/PyResult
- M-5: aggregates civ_id UInt16->UInt8
- belief column restored in events schema/batch builder
- civ_id column restored in promotions schema/batch builder
- Martyrdom guard for persecution_intensity and minority belief
- _apply_civ_op re-exported from politics.py
- Pool test downcast types fixed for UInt8/UInt16 consistency

+++ chronicler-agents/src/pool.rs
@@ -1151,7 +1151,7 @@ mod tests {
     fn test_compute_aggregates_zero_agents_with_controlled_region() {
         // M-AF1 #7: A civ controls a region but has zero alive agents.
         // compute_aggregates must emit a zero-row for that civ.
-        use arrow::array::{UInt16Array, UInt32Array};
+        use arrow::array::{UInt32Array, UInt8Array};
         use crate::region::RegionState;
 
         let pool = AgentPool::new(4); // No agents spawned at all.
@@ -1171,7 +1171,7 @@ mod tests {
         let civ_ids = batch
             .column(0)
             .as_any()
-            .downcast_ref::<UInt16Array>()
+            .downcast_ref::<UInt8Array>()
             .unwrap();
         assert_eq!(civ_ids.value(0), 0);
 
@@ -1197,7 +1197,7 @@ mod tests {
     fn test_compute_aggregates_mixed_zero_and_populated_civs() {
         // M-AF1 #7: civ 0 has agents, civ 1 controls a region but has no agents.
         // Both must appear in the output.
-        use arrow::array::{UInt16Array, UInt32Array};
+        use arrow::array::{UInt32Array, UInt8Array};
         use crate::region::RegionState;
 
         let mut pool = AgentPool::new(8);
@@ -1223,7 +1223,7 @@ mod tests {
         let civ_ids = batch
             .column(0)
             .as_any()
-            .downcast_ref::<UInt16Array>()
+            .downcast_ref::<UInt8Array>()
             .unwrap();
         let populations = batch
             .column(1)
